Write out “JavaScript” rather than “JS”
Casually-abbreviated “JS” in “JS runtimes” isn’t strictly wrong, but
“JavaScript runtimes” — written out — is better in this context (which
is effectively a standalone mission statement) because it’s completely
clear and unambiguous. Otherwise, since there’s nowhere else any
explicit mention of JavaScript in the page, using just “JS” risks being
less clear to some readers. And as a kind of mission statement, it’s
better for the language here to be a bit more formal and less casual.